server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.9.2
pydantic-settings==2.5.2
supabase==2.9.0
boto3==1.35.36
httpx==0.27.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
pytest==8.3.3
pytest-asyncio==0.24.0
